product filter added

See merge request ozod_saidov/sprinter!18

+++ core/models/product.py
@@ -35,7 +35,7 @@ class Product(Base):
     title = models.CharField(max_length=255)
     description = models.TextField()
     price = models.FloatField()
-    discount = models.FloatField('Скидка', null=True, blank=True)
+    discount = models.PositiveSmallIntegerField('Скидка', null=True, blank=True)
     old_price = models.FloatField(null=True, blank=True)
     is_active = models.BooleanField(default=True)
     available_quantity = models.PositiveIntegerField(default=0)
